chore: moves log to top-level package
